Fix failing controller specs by stubbing refinery_user_required?.
In Refinery CMS 2.0.x we require user with refinery role to be present
even if we visit only frontend pages. In Refinery CMS 2.1.x this check
was changed to apply only to backend stuff so in order for specs to pass
we need to a) create refinery user or b) stub refinery_user_required?
and return false.